+++ Chatbot_Admin/src/main/java/com/psib/service/IRoleManager.java
@@ -11,5 +11,5 @@ import com.psib.model.Role;
  * @Email: datht0601@gmail.com
  */
 public interface IRoleManager {
-	public Role getRoleById(long id);
+	public Role getRoleById(int id);
 }
